+++ packages/guides-graphs/resources/config/guides-graphs.php
@@ -3,12 +3,10 @@
 declare(strict_types=1);
 
 use phpDocumentor\Guides\Graphs\Directives\UmlDirective;
-use phpDocumentor\Guides\Graphs\Nodes\UmlNode;
 use phpDocumentor\Guides\Graphs\Renderer\DiagramRenderer;
 use phpDocumentor\Guides\Graphs\Renderer\PlantumlRenderer;
 use phpDocumentor\Guides\Graphs\Renderer\PlantumlServerRenderer;
 use phpDocumentor\Guides\Graphs\Twig\UmlExtension;
-use phpDocumentor\Guides\NodeRenderers\TemplateNodeRenderer;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 
 return static function (ContainerConfigurator $container): void {
@@ -18,11 +16,6 @@ return static function (ContainerConfigurator $container): void {
         ->autoconfigure()
         ->set(UmlDirective::class)
         ->tag('phpdoc.guides.directive')
-
-        ->set('phpdoc.guides.', TemplateNodeRenderer::class)
-        ->tag('phpdoc.guides.noderenderer.html')
-        ->arg('$template', 'body/uml.html.twig')
-        ->arg('$nodeClass', UmlNode::class)
 
         ->set(PlantumlRenderer::class)
         ->arg('$plantUmlBinaryPath', '%guides.graphs.plantuml_binary%')
