+++ packages/guides-restructured-text/src/RestructuredText/Directives/ConfvalDirective.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace phpDocumentor\Guides\RestructuredText\Directives;
 
+use phpDocumentor\Guides\NodeRenderers\Html\GeneralNodeHtmlRenderer;
+use phpDocumentor\Guides\NodeRenderers\LaTeX\GeneralNodeLatexRenderer;
 use phpDocumentor\Guides\Nodes\CollectionNode;
 use phpDocumentor\Guides\Nodes\Node;
 use phpDocumentor\Guides\ReferenceResolvers\AnchorReducer;
@@ -30,11 +32,15 @@ final class ConfvalDirective extends SubDirective
     public function __construct(
         protected Rule $startingRule,
         GenericLinkProvider $genericLinkProvider,
+        GeneralNodeHtmlRenderer $generalNodeRenderer,
+        GeneralNodeLatexRenderer $generalNodeLatexRenderer,
         private readonly AnchorReducer $anchorReducer,
         private readonly InlineParser $inlineParser,
     ) {
         parent::__construct($startingRule);
 
+        $generalNodeRenderer->registerNode(ConfvalNode::class, 'body/directive/confval.html.twig');
+        $generalNodeLatexRenderer->registerNode(ConfvalNode::class, 'body/directive/confval.tex.twig');
         $genericLinkProvider->addGenericLink(self::NAME, ConfvalNode::LINK_TYPE);
     }
