+++ packages/guides-restructured-text/src/RestructuredText/Directives/VersionChangedDirective.php
@@ -13,12 +13,18 @@ declare(strict_types=1);
 
 namespace phpDocumentor\Guides\RestructuredText\Directives;
 
+use phpDocumentor\Guides\NodeRenderers\Html\GeneralNodeHtmlRenderer;
+use phpDocumentor\Guides\RestructuredText\Nodes\VersionChangeNode;
 use phpDocumentor\Guides\RestructuredText\Parser\Productions\Rule;
 
 class VersionChangedDirective extends AbstractVersionChangeDirective
 {
-    public function __construct(protected Rule $startingRule)
-    {
+    public function __construct(
+        protected Rule $startingRule,
+        GeneralNodeHtmlRenderer $generalNodeRenderer,
+    ) {
         parent::__construct($startingRule, 'versionchanged', 'Changed in version %s');
+
+        $generalNodeRenderer->registerNode(VersionChangeNode::class, 'body/version-change.html.twig');
     }
 }
