+++ packages/guides/resources/template/tex/inline/textroles/generic.tex.twig
@@ -1,0 +1,4 @@
+{%- include  [
+    ('inline/textroles/' ~ node.type ~ '.tex.twig'),
+    'inline/textroles/unkown.tex.twig'
+    ]  -%}
