+++ packages/guides/resources/template/tex/inline/textroles/unkown.tex.twig
@@ -1,0 +1,3 @@
+{% apply spaceless %}
+\texttt{{ '{' }}{{- node.value -}}{{ '}' }}
+{% endapply %}
