+++ packages/guides/src/NodeRenderers/Html/GeneralNodeHtmlRenderer.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace phpDocumentor\Guides\NodeRenderers\Html;
+
+use phpDocumentor\Guides\NodeRenderers\GeneralNodeRenderer;
+
+final class GeneralNodeHtmlRenderer extends GeneralNodeRenderer
+{
+}
