+++ packages/guides/src/NodeRenderers/LaTeX/GeneralNodeLatexRenderer.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace phpDocumentor\Guides\NodeRenderers\LaTeX;
+
+use phpDocumentor\Guides\NodeRenderers\GeneralNodeRenderer;
+
+final class GeneralNodeLatexRenderer extends GeneralNodeRenderer
+{
+}
