+++ tests/Integration/tests/latex/latex-confval/expected/index.tex
@@ -1,0 +1,28 @@
+\documentclass[11pt]{report}
+\usepackage[utf8]{inputenc}
+\usepackage[T1]{fontenc}
+\usepackage[french]{babel}
+\usepackage{cite}
+\usepackage{amssymb}
+\usepackage{amsmath}
+\usepackage{mathrsfs}
+\usepackage{graphicx}
+\usepackage{hyperref}
+\usepackage{listings}
+\begin{document}
+\label{}
+    \section{Confval directive}
+
+    \begin{description}
+    \item[\textttdemo]
+    \hfill \\
+    \begin{itemize}
+        \item[\textbf{Type:}] string
+        \item[\textbf{Required:}] false        \item[\textbf{Default:}] N/A\end{itemize}
+
+    \item[Description]
+    \hfill \\
+This is the confval \texttt{demo} content!
+\end{description}
+
+\end{document}
