+++ tests/Integration/tests/latex/latex-confval/input/index.rst
@@ -1,0 +1,7 @@
+Confval directive
+=================
+
+.. confval:: demo
+    :type: string
+
+    This is the confval `demo` content!
